+++ packages/terra-infinite-list/CHANGELOG.md
@@ -8,6 +8,7 @@
 ## 3.34.0 - (December 8, 2020)
 
 * Changed
+  * Updated to support a peer dependency of react-intl v2-v5
   * Update wdio tests to remove deprecated Terra.it helpers
 
 ## 3.33.0 - (November 10, 2020)