+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -18,8 +18,9 @@
 ## 4.4.0 - (November 10, 2020)
 
 * Changed
+  * Updated to support a peer dependency of react-intl v2-v5
   * Updated jest  due to changes in abstract modal.
-  
+
 ## 4.3.0 - (October 20, 2020)
 
 * Changed